+++ backend/settings.php
@@ -244,7 +244,7 @@
       <ul>
         <li><a href="manageStaticPages.php">Manage Static Pages</a></li>
         <li><a href="managePrivileges.php">User &amp; Role Privileges</a></li>
-        <li><a href="index.php">Backend Dashboard</a></li>
+        <li><a href="index.php">Admin Hub</a></li>
       </ul>
     </div>
   </div>